refactor: keep sidebar task components under limits
